requirements.txt
absl-py==0.8.0
aniso8601==8.0.0
ansi2html==1.5.2
argh==0.26.2
arrow==0.15.5
asn1crypto==0.24.0
astor==0.8.0
asyncio==3.4.3
attrs==19.3.0
audioread==2.1.8
backcall==0.1.0
bleach==3.1.0
boto3==1.11.7
botocore==1.14.7
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cmake==3.13.3
colorama==0.4.3
configparser==4.0.2
cryptography==2.1.4
cycler==0.10.0
Cython==0.29.14
decorator==4.4.1
defusedxml==0.6.0
dill==0.3.1.1
dlib==19.19.0
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.15.2
entrypoints==0.3
ffmpeg==1.4
Flask==1.1.1
Flask-RESTful==0.3.7
funcsigs==1.0.2
future==0.18.2
gast==0.2.2
gitdb2==2.0.6
GitPython==3.0.5
google-pasta==0.1.7
gql==0.2.0
graphql-core==1.1
grpcio==1.24.0
h5py==2.10.0
idna==2.6
image-classifiers==1.0.0
imageio==2.6.1
imageio-ffmpeg==0.3.0
imgaug==0.3.0
importlib-metadata==1.4.0
ipykernel==5.1.3
ipython==7.11.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.15.2
Jinja2==2.10.3
jiwer==1.3.2
jmespath==0.9.4
joblib==0.14.1
json5==0.8.5
jsonpickle==1.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
librosa==0.7.2
llvmlite==0.31.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mir-eval==0.5
mistune==0.8.4
more-itertools==8.1.0
moviepy==1.0.1
mtcnn==0.1.0
munch==2.5.0
namedtupled==0.3.3
natsort==7.0.0
nbconvert==5.6.1
nbformat==5.0.3
networkx==2.4
nltk==3.4.5
notebook==6.0.3
numba==0.47.0
numpy==1.17.2
nvgpu==0.8.0
nvidia-ml-py3==7.352.0
opencv-python==4.1.2.30
opencv-python-headless==4.1.2.30
opt-einsum==3.1.0
packaging==20.1
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.2
pathlib==1.0.1
pathtools==0.1.2
pesq==0.0.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
pipreqs==0.4.10
proglog==0.1.9
prometheus-client==0.7.1
promise==2.3
prompt-toolkit==3.0.2
protobuf==3.9.2
psutil==5.6.7
ptyprocess==0.6.0
py-cpuinfo==5.0.0
pycparser==2.19
pycrypto==2.6.1
pydub==0.23.1
pyemd==0.5.1
Pygments==2.5.2
pygobject==3.26.1
pygraphviz==1.5
pymongo==3.10.1
pynvml==8.0.4
pyparsing==2.4.6
pyrsistent==0.15.7
pystoi==0.2.2
python-apt==1.6.4
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
pyxdg==0.25
PyYAML==5.3
pyzmq==18.1.1
qtconsole==4.6.0
regex==2020.1.8
requests==2.22.0
resampy==0.2.2
s3transfer==0.3.1
sacred==0.8.1
sacremoses==0.0.38
scikit-image==0.16.2
scikit-learn==0.22.1
scikit-video==1.1.11
scipy==1.4.1
seaborn==0.9.0
SecretStorage==2.3.1
Send2Trash==1.5.0
sentencepiece==0.1.85
sentry-sdk==0.14.1
seqeval==0.0.5
Shapely==1.6.4.post2
shortuuid==0.5.0
six==1.11.0
sk-video==1.1.10
smmap2==2.0.5
SoundFile==0.10.3.post1
subprocess32==3.5.4
tabulate==0.8.6
tensorboard==2.0.0
tensorflow-estimator==2.0.0
tensorflow-gpu==2.0.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
torch==1.4.0
torchvision==0.5.0
tornado==6.0.3
tqdm==4.41.1
traitlets==4.3.3
transformers==2.3.0
urllib3==1.25.8
wandb==0.8.21
watchdog==0.9.0
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
yarg==0.1.9
zipp==2.0.0
